vk_echo_bot/requirements.txt
vkbottle==4.3.12
python-dotenv==1.0.0
